Refactor serializers of the pipeline
